+++ R/req-perform.R
@@ -59,8 +59,8 @@
 #'   * If the HTTP request fails (e.g. the connection is dropped or the
 #'     server doesn't exist), an error with class `"httr2_failure"`.
 #' @export
-#' @seealso [req_perform_multi()] to perform multiple requests in parallel.
-#'   [req_perform_paginate()] to fetch all pages of a requests paginated via
+#' @seealso [req_perform_parallel()] to perform multiple requests in parallel.
+#'   [req_perform_iterate()] to fetch all pages of a requests paginated via
 #'   [req_paginate()].
 #' @examples
 #' request("https://google.com") %>%