+++ tests/testthat/test-multi-req.R
@@ -1,6 +1,6 @@
 test_that("request and paths must match", {
   req <- request("http://example.com")
-  expect_snapshot(req_perform_multi(req, letters), error = TRUE)
+  expect_snapshot(req_perform_parallel(req, letters), error = TRUE)
 })
 
 test_that("requests happen in parallel", {
@@ -17,14 +17,14 @@ test_that("requests happen in parallel", {
     request_test("/delay/:secs", secs = 0.25),
     request_test("/delay/:secs", secs = 0.25),
   )
-  time <- system.time(req_perform_multi(reqs))
+  time <- system.time(req_perform_parallel(reqs))
   expect_lt(time[[3]], 1)
 })
 
 test_that("can download files", {
   reqs <- list(request_test("/json"), request_test("/html"))
   paths <- c(withr::local_tempfile(), withr::local_tempfile())
-  resps <- req_perform_multi(reqs, paths)
+  resps <- req_perform_parallel(reqs, paths)
 
   expect_equal(resps[[1]]$body, new_path(paths[[1]]))
   expect_equal(resps[[2]]$body, new_path(paths[[2]]))
@@ -43,7 +43,7 @@ test_that("immutable objects retrieved from cache", {
   cache_set(req, resp)
 
   expect_condition(
-    resps <- req_perform_multi(list(req)),
+    resps <- req_perform_parallel(list(req)),
     class = "httr2_cache_cached"
   )
   expect_equal(resps[[1]], resp)
@@ -54,7 +54,7 @@ test_that("both curl and HTTP errors become errors", {
     request_test("/status/:status", status = 404),
     request("INVALID"),
   )
-  out <- req_perform_multi(reqs)
+  out <- req_perform_parallel(reqs)
   expect_s3_class(out[[1]], "httr2_http_404")
   expect_s3_class(out[[2]], "httr2_failure")
 })
@@ -64,7 +64,7 @@ test_that("errors can cancel outstanding requests", {
     request_test("/status/:status", status = 404),
     request_test("/delay/:secs", secs = 2),
   )
-  out <- req_perform_multi(reqs, cancel_on_error = TRUE)
+  out <- req_perform_parallel(reqs, cancel_on_error = TRUE)
   expect_s3_class(out[[1]], "httr2_http_404")
   expect_s3_class(out[[2]], "httr2_cancelled")
 
@@ -72,7 +72,7 @@ test_that("errors can cancel outstanding requests", {
     request("blah://INVALID"),
     request_test("/delay/:secs", secs = 2),
   )
-  out <- req_perform_multi(reqs, cancel_on_error = TRUE)
+  out <- req_perform_parallel(reqs, cancel_on_error = TRUE)
   expect_s3_class(out[[1]], "httr2_failure")
   expect_s3_class(out[[2]], "httr2_cancelled")
 })